py/packages/genkit/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# SPDX-License-Identifier: Apache-2.0

[project]
authors = [{ name = "Google" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Web Environment",
  "Framework :: AsyncIO",
  "Framework :: Pydantic",
  "Framework :: Pydantic :: 2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries",
  "Typing :: Typed",
  "License :: OSI Approved :: Apache Software License",
]
dependencies = [
  "opentelemetry-api>=1.29.0",
  "opentelemetry-sdk>=1.29.0",
  "pydantic>=2.10.5",
  "partial-json-parser>=0.2.1.1.post5",
  "json5>=0.10.0",
  "structlog>=25.2.0",
  "rich>=13.0.0",
  "asgiref>=3.8.1",
  "httpx>=0.28.1",
  "psutil>=7.0.0",
  "starlette>=0.46.1",
  "python-multipart>=0.0.22",
  "sse-starlette>=2.2.1",
  "pillow>=12.1.1",
  "typing_extensions>=4.0",
  "strenum>=0.4.15; python_version < '3.11'",
  "dotpromptz>=0.1.5",
  "uvicorn>=0.34.0",
  "uvloop>=0.21.0; sys_platform != 'win32'",
  "anyio>=4.9.0",
  "opentelemetry-instrumentation-logging>=0.60b1",
]
description = "Genkit AI Framework"
keywords = [
  "genkit",
  "ai",
  "llm",
  "machine-learning",
  "artificial-intelligence",
  "generative-ai",
  "framework",
  "sdk",
]
license = "Apache-2.0"
name = "genkit"
readme = "README.md"
requires-python = ">=3.10"
version = "0.5.1"

[project.optional-dependencies]
dev-local-vectorstore = ["genkit-plugin-dev-local-vectorstore"]
flask                 = ["genkit-plugin-flask"]
google-cloud          = ["genkit-plugin-google-cloud"]
google-genai          = ["genkit-plugin-google-genai"]
ollama                = ["genkit-plugin-ollama"]
openai                = ["genkit-plugin-compat-oai"]
vertex-ai             = ["genkit-plugin-vertex-ai"]

[project.urls]
"Bug Tracker"   = "https://github.com/firebase/genkit/issues"
"Changelog"     = "https://github.com/firebase/genkit/blob/main/py/packages/genkit/CHANGELOG.md"
"Documentation" = "https://firebase.google.com/docs/genkit"
"Homepage"      = "https://github.com/firebase/genkit"
"Repository"    = "https://github.com/firebase/genkit/tree/main/py"

[build-system]
build-backend = "hatchling.build"
requires      = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/genkit"]

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests", "packages/genkit/src"]
testpaths  = ["tests"]

[tool.uv.sources]
genkit-plugin-compat-oai            = { workspace = true }
genkit-plugin-dev-local-vectorstore = { workspace = true }
genkit-plugin-flask                 = { workspace = true }
genkit-plugin-google-genai          = { workspace = true }
genkit-plugin-ollama                = { workspace = true }
